requirements.txt
Django==4.*
amqp
billiard
bleach
bootstrap-admin

certifi
chardet
defusedxml

django-appconf
django-crispy-forms
django-haystack
django-imagekit
django-redis
django-tctip
django-timezone-field
djangorestframework
django-allauth
gunicorn
html5lib

idna
importlib-metadata
jieba
kombu
Markdown
markdown-checklist
mysqlclient
oauthlib
olefile
packaging
pilkit
Pillow==9.3.0
Pygments
PyMySQL
pyparsing
python-crontab
python-dateutil
python3-openid
pytz
redis
requests
requests-oauthlib
setuptools-scm
six
sqlparse
supervisor
urllib3
user-agent
vine
webencodings
Whoosh
zipp
python-dotenv